// apps/admin/src/lib/api/generated/models/appearanceSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.4
 */
import type { AppearanceSettingsDateFormat } from './appearanceSettingsDateFormat';
import type { AppearanceSettingsDensity } from './appearanceSettingsDensity';
import type { AppearanceSettingsTheme } from './appearanceSettingsTheme';
import type { AppearanceSettingsTimeFormat } from './appearanceSettingsTimeFormat';

export interface AppearanceSettings {
  date_format?: AppearanceSettingsDateFormat;
  density?: AppearanceSettingsDensity;
  reduce_motion?: boolean;
  theme?: AppearanceSettingsTheme;
  time_format?: AppearanceSettingsTimeFormat;
  /** @maxLength 100 */
  timezone?: string;
}