// apps/admin/src/lib/api/generated/models/defaultSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.7
 */
import type { DefaultSettingsLandingPage } from './defaultSettingsLandingPage';
import type { DefaultSettingsOverviewDays } from './defaultSettingsOverviewDays';
import type { DefaultSettingsPageSize } from './defaultSettingsPageSize';
import type { DefaultSettingsRefreshSeconds } from './defaultSettingsRefreshSeconds';

export interface DefaultSettings {
  landing_page?: DefaultSettingsLandingPage;
  overview_days?: DefaultSettingsOverviewDays;
  page_size?: DefaultSettingsPageSize;
  refresh_seconds?: DefaultSettingsRefreshSeconds;
  remember_columns?: boolean;
  remember_filters?: boolean;
  remember_sort?: boolean;
}